checkpoint: riorganizzazione cartella, documento master, knowledge graph
- Struttura source riorganizzata in moduli: Views/, Domain/, Models/,
  Auth/, Export/, Intents/, Debug/, Persistence/, Formatting/,
  OnboardingTour/, Sync/
- Eliminati file sparsi in root (audit vecchi, .Rhistory, worktree stantio)
- Creato docs/MoneyTracker.md — documento unico con visione, architettura,
  funzionalità, audit, roadmap, changelog, setup
- Generato docs/MoneyTracker.pdf (454 KB)
- Aggiunto CLAUDE.md con workflow checkpoint
- Costruito knowledge graph graphify-out/ (700 nodi, 1346 archi, 35 community)

+++ MoneyTracker/Domain/NotificationScheduling.swift
@@ -20,7 +20,7 @@ protocol NotificationScheduling: AnyObject {
     func scheduleGoalDeadline(name: String, goalId: String, deadline: Date)
     func cancelGoalDeadline(goalId: String)
 
-    func checkBalanceThreshold(accountId: String, accountName: String, currentBalance: Double)
+    func checkBalanceThreshold(accountId: String, accountName: String, currentBalance: Decimal)
 
     func scheduleMonthlyReportReminder(enabled: Bool)
 }
